# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce-BackEnd-Documentation",
        "description": "Docuementation for Ecommerce-BackEnd",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8017",
            "description": "Develop Server"
        }
    ],
    "paths": {
...
        "/api/reservation/getreservation": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Obtener todos las reservaciones",
                "description": "Este EndPoint regresa todos las reservaciones.",
                "operationId": "allreservation",
                "responses": {
                    "200": {
                        "description": "Todos las reservaciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "phone_number": {
                                                        "type": "string",
                                                        "example": "1234567899"
                                                    },
                                                    "skype": {
                                                        "type": "string",
                                                        "example": "example.skype"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-12-12"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2024-12-12"
                                                    },
                                                    "deleted_at": {
                                                        "type": "string",
                                                        "example": "null"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reservaciones obtenidas correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error del servidor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reservation/getReservation/{id}": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Obtener una reservacion",
                "description": "Este EndPoint regresa una reservacion por su ID.",
                "operationId": "getreservation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reservacion",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservacion obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "phone_number": {
                                                    "type": "string",
                                                    "example": "1234567899"
                                                },
                                                "skype": {
                                                    "type": "string",
                                                    "example": "example.skype"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-12-12"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-12-12"
                                                },
                                                "deleted_at": {
                                                    "type": "string",
                                                    "example": "null"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reservacion obtenido correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Medio no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reservacion no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error del servidor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}